+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/tool/MCPToolsScreen.kt
@@ -4,22 +4,22 @@ import ai.mcpdirect.studio.app.auth.authViewModel
 import ai.mcpdirect.studio.app.compose.StudioCard
 import ai.mcpdirect.studio.app.compose.StudioIcon
 import ai.mcpdirect.studio.app.compose.StudioListItem
-import ai.mcpdirect.studio.app.compose.Tag
+import ai.mcpdirect.studio.app.compose.TooltipIconButton
 import ai.mcpdirect.studio.app.compose.TooltipText
 import ai.mcpdirect.studio.app.generalViewModel
 import ai.mcpdirect.studio.app.model.account.AIPortTeam
+import ai.mcpdirect.studio.app.model.aitool.AIPortToolAgent
 import ai.mcpdirect.studio.app.model.aitool.AIPortToolMaker
 import ai.mcpdirect.studio.app.template.CreateMCPServerTemplateDialog
-import ai.mcpdirect.studio.app.template.mcpTemplateViewModel
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import mcpdirectstudioapp.composeapp.generated.resources.Res
@@ -27,6 +27,7 @@ import mcpdirectstudioapp.composeapp.generated.resources.diversity_3
 import mcpdirectstudioapp.composeapp.generated.resources.graph_5
 import mcpdirectstudioapp.composeapp.generated.resources.keyboard_arrow_right
 import mcpdirectstudioapp.composeapp.generated.resources.person
+import mcpdirectstudioapp.composeapp.generated.resources.share
 import org.jetbrains.compose.resources.painterResource
 
 private enum class MCPToolsScreenDialog {
@@ -38,8 +39,9 @@ private var dialog by mutableStateOf(MCPToolsScreenDialog.NONE)
 @Composable
 fun MCPToolsScreen() {
     LaunchedEffect(null){
-        generalViewModel.refreshToolMakers()
+        generalViewModel.refreshToolAgents()
         generalViewModel.refreshTeams()
+        generalViewModel.topBarActions = {}
     }
     when(dialog){
         MCPToolsScreenDialog.NONE -> {}
@@ -50,8 +52,9 @@ fun MCPToolsScreen() {
                     CreateMCPServerTemplateDialog(
                         toolMaker,
                         config = it,
-                        onConfirmRequest = {
+                        onConfirmRequest = { name,type,agentId,config,inputs ->
                             dialog = MCPToolsScreenDialog.NONE
+                            mcpToolsViewModel.createToolMakerTemplate(name,type,agentId,config,inputs)
                         },
                         onDismissRequest = {
                             dialog = MCPToolsScreenDialog.NONE
@@ -70,7 +73,10 @@ fun MCPToolsScreen() {
                 tabs.forEachIndexed { index, title ->
                     Tab(
                         selected = currentTabIndex == index,
-                        onClick = { currentTabIndex = index },
+                        onClick = {
+                            currentTabIndex = index
+                            generalViewModel.topBarActions = {}
+                                  },
                         text = { Text(title) }
                     )
                 }
@@ -79,7 +85,7 @@ fun MCPToolsScreen() {
             // Content based on selected tab
             when (currentTabIndex) {
                 0 -> MCPServerList()
-                1 -> {}
+                1 -> MCPTemplateList()
             }
 
         }
@@ -88,7 +94,7 @@ fun MCPToolsScreen() {
         StudioCard(Modifier.fillMaxSize().padding(8.dp).weight(2.0f)) {
             when (currentTabIndex) {
                 0 -> MCPServerItem()
-                1 -> {}
+                1 -> MCPTemplateItem()
             }
         }
     }
@@ -96,6 +102,9 @@ fun MCPToolsScreen() {
 
 @Composable
 fun MCPServerList(){
+    LaunchedEffect(null){
+        generalViewModel.refreshToolMakers()
+    }
     val viewModel = mcpToolsViewModel
     LazyColumn {
         items(generalViewModel.toolMakers){
@@ -136,7 +145,7 @@ fun MCPServerList(){
                         Text(
                             if(it.type== AIPortToolMaker.TYPE_VIRTUAL)
                                 "Virtual MCP"
-                            else it.agentName ?:"",
+                            else it.agentName,
                             color = MaterialTheme.colorScheme.primary,
                             style=MaterialTheme.typography.bodySmall
                         )
@@ -189,6 +198,100 @@ fun MCPServerItem(){
             Row {
 
             }
+        }
+    }
+}
+
+@Composable
+fun MCPTemplateList(){
+    val viewModel = remember { mcpToolsViewModel }
+    LaunchedEffect(viewModel){
+        viewModel.queryToolMakerTemplates()
+    }
+    LazyColumn {
+        items(viewModel.toolMakerTemplates){ template ->
+            val me = template.userId== authViewModel.user.id
+            var toolAgent by remember { mutableStateOf<AIPortToolAgent?>(null) }
+            LaunchedEffect(null){
+                generalViewModel.toolAgent(template.agentId){
+                    code, message, data ->
+                    toolAgent = data
+                }
+            }
+            toolAgent?.let { agent ->
+                StudioListItem(
+                    selected = viewModel.toolMakerTemplate?.id == template.id,
+                    modifier = Modifier.clickable{
+                        viewModel.toolMakerTemplate(template)
+                    },
+                overlineContent = {
+                    Row {
+                        StudioIcon(
+                            Res.drawable.person,
+                            "Owner",
+                            modifier = Modifier.size(18.dp).padding(end = 4.dp)
+                        )
+                        if(me) Text("Me")
+//                        else if(team!=null) TooltipText(
+//                            team.ownerName,
+//                            contentDescription = team.ownerAccount,
+//                            overflow = TextOverflow.Ellipsis
+//                        )
+                    }
+                },
+                    headlineContent = { Text(
+                        template.name,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis) },
+                    supportingContent = {
+                        Row{
+                            StudioIcon(
+                                Res.drawable.graph_5,
+                                "From My Studio",
+                                MaterialTheme.colorScheme.primary,
+                                modifier = Modifier.size(20.dp).padding(end = 4.dp)
+                            )
+                            Text(
+                                agent.name,
+                                color = MaterialTheme.colorScheme.primary,
+                                style=MaterialTheme.typography.bodySmall
+                            )
+
+                        }
+
+                    },
+                    trailingContent = {
+                        if(viewModel.toolMakerTemplate!=null&&viewModel.toolMakerTemplate!!.id==template.id)
+                            Icon(painterResource(Res.drawable.keyboard_arrow_right),
+                                contentDescription = "")
+                    },
+                    colors = ListItemDefaults.colors(
+                        containerColor = if(viewModel.toolMakerTemplate==template)
+                            MaterialTheme.colorScheme.surfaceContainer
+                        else Color.Transparent
+                    )
+                )
+                HorizontalDivider()
+            }
+        }
+    }
+}
+
+@Composable
+fun MCPTemplateItem(){
+    mcpToolsViewModel.toolMakerTemplate?.let {
+        Column {
+            Row(Modifier.padding(start = 8.dp, end = 8.dp),
+                verticalAlignment = Alignment.CenterVertically){
+                Text(it.name)
+                Spacer(Modifier.weight(1.0f))
+                TooltipIconButton(
+                    Res.drawable.share,
+                    "Share MCP Template to Team",
+                    onClick = {}
+                )
+            }
+            HorizontalDivider()
         }
     }
 }